+++ source/modules/database/asp/index.js
@@ -168,6 +168,13 @@ class ASP {
       type: 'button',
       icon: 'remove',
       text: LANG['form']['toolbar']['clear']
+    }, {
+      type: 'separator'
+    }, {
+      id: 'test',
+      type: 'button',
+      icon: 'spinner',
+      text: LANG['form']['toolbar']['test']
     }]);
 
     // form
@@ -225,6 +232,32 @@ class ASP {
             this.manager.list.imgs[0]
           );
           break;
+        case 'test':
+          if (!form.validate()) {
+            return toastr.warning(LANG['form']['warning'], LANG_T['warning']);
+          };
+          // 解析数据
+          let _data = form.getValues();
+          win.progressOn();
+          this.core.request(
+            this.core[`database_${_data['type']}`].show_databases({
+              conn: _data['conn']
+            })
+          ).then((res) => {
+            if(res['text'].length > 0){
+              if(res['text'].indexOf("ERROR://") > -1) {
+                throw res["text"];
+              }
+              toastr.success(LANG['form']['test_success'], LANG_T['success']);
+            }else{
+              toastr.warning(LANG['form']['test_warning'], LANG_T['warning']);
+            }
+            win.progressOff();
+          }).catch((err)=>{
+            win.progressOff();
+            toastr.error(JSON.stringify(err), LANG_T['error']);
+          });
+          break;
       }
     });
   }
@@ -259,6 +292,13 @@ class ASP {
       type: 'button',
       icon: 'remove',
       text: LANG['form']['toolbar']['clear']
+    }, {
+      type: 'separator'
+    }, {
+      id: 'test',
+      type: 'button',
+      icon: 'spinner',
+      text: LANG['form']['toolbar']['test']
     }]);
 
     // form
@@ -309,6 +349,32 @@ class ASP {
           toastr.success(LANG['form']['success'], LANG_T['success']);
           // 刷新 UI
           this.parse();
+          break;
+        case 'test':
+          if (!form.validate()) {
+            return toastr.warning(LANG['form']['warning'], LANG_T['warning']);
+          };
+          // 解析数据
+          let _data = form.getValues();
+          win.progressOn();
+          this.core.request(
+            this.core[`database_${_data['type']}`].show_databases({
+              conn: _data['conn']
+            })
+          ).then((res) => {
+            if(res['text'].length > 0){
+              if(res['text'].indexOf("ERROR://") > -1) {
+                throw res["text"];
+              }
+              toastr.success(LANG['form']['test_success'], LANG_T['success']);
+            }else{
+              toastr.warning(LANG['form']['test_warning'], LANG_T['warning']);
+            }
+            win.progressOff();
+          }).catch((err)=>{
+            win.progressOff();
+            toastr.error(JSON.stringify(err), LANG_T['error']);
+          });
           break;
       }
     });
@@ -438,7 +504,7 @@ class ASP {
       this.core[`database_${conf['type']}`].show_columns(
       {
         conn: conf['conn'],
-        table: conf['type'] === 'oracle' ? `SELECT * FROM (SELECT A.*,ROWNUM N FROM ${table} A) WHERE N=1` : `SELECT TOP 1 * FROM ${table}`
+        table: conf['type'] === 'oracle' ? `SELECT * FROM (SELECT A.*,ROWNUM N FROM ${table} A) WHERE N=1` : `USE [${this.dbconf['database']}];SELECT TOP 0 * FROM ${table}`
       })
     ).then((res) => {
       let ret = res['text'];
@@ -479,7 +545,8 @@ class ASP {
     this.core.request(
       this.core[`database_${this.dbconf['type']}`].query({
         conn: this.dbconf['conn'],
-        sql: sql
+        sql: sql,
+        dbname: this.dbconf['database'],
       })
     ).then((res) => {
       let ret = res['text'];